+++ docs/conf.py
@@ -32,7 +32,7 @@ import git
 # ones.
 extensions = ['sphinxcontrib.httpdomain', 'sphinx.ext.extlinks']
 
-extlinks = {'repo': ('https://github.com/lyft/envoy/blob/master/%s', '')}
+extlinks = {'repo': ('https://github.com/envoyproxy/envoy/blob/master/%s', '')}
 
 # Add any paths that contain templates here, relative to this directory.
 templates_path = ['_templates']
@@ -139,8 +139,7 @@ html_theme_path = [sphinx_rtd_theme.get_html_theme_path()]
 # The name of an image file (relative to this directory) to use as a favicon of
 # the docs.  This file should be a Windows icon file (.ico) being 16x16 or 32x32
 # pixels large.
-# TODO(mattklein123): Restore this.
-#html_favicon = 'landing_source/source/favicon.ico'
+html_favicon = 'favicon.ico'
 
 # Add any paths that contain custom static files (such as style sheets) here,
 # relative to this directory. They are copied after the builtin static files,