+++ include/envoy/router/router.h
@@ -11,6 +11,7 @@
 #include "envoy/common/optional.h"
 #include "envoy/http/codec.h"
 #include "envoy/http/header_map.h"
+#include "envoy/tracing/http_tracer.h"
 #include "envoy/upstream/resource_manager.h"
 
 namespace Envoy {
@@ -29,6 +30,49 @@ public:
    * @return std::string the redirect URL.
    */
   virtual std::string newPath(const Http::HeaderMap& headers) const PURE;
+};
+
+/**
+ * CorsPolicy for Route and VirtualHost.
+ */
+class CorsPolicy {
+public:
+  virtual ~CorsPolicy() {}
+
+  /**
+   * @return std::list<std::string>& access-control-allow-origin values.
+   */
+  virtual const std::list<std::string>& allowOrigins() const PURE;
+
+  /**
+   * @return std::string access-control-allow-methods value.
+   */
+  virtual const std::string& allowMethods() const PURE;
+
+  /**
+   * @return std::string access-control-allow-headers value.
+   */
+  virtual const std::string& allowHeaders() const PURE;
+
+  /**
+   * @return std::string access-control-expose-headers value.
+   */
+  virtual const std::string& exposeHeaders() const PURE;
+
+  /**
+   * @return std::string access-control-max-age value.
+   */
+  virtual const std::string& maxAge() const PURE;
+
+  /**
+   * @return const Optional<bool>& Whether access-control-allow-credentials should be true.
+   */
+  virtual const Optional<bool>& allowCredentials() const PURE;
+
+  /**
+   * @return bool Whether CORS is enabled for the route or virtual host.
+   */
+  virtual bool enabled() const PURE;
 };
 
 /**
@@ -147,6 +191,11 @@ public:
   virtual ~VirtualHost() {}
 
   /**
+   * @return const CorsPolicy* the CORS policy for this virtual host.
+   */
+  virtual const CorsPolicy* corsPolicy() const PURE;
+
+  /**
    * @return const std::string& the name of the virtual host.
    */
   virtual const std::string& name() const PURE;
@@ -185,6 +234,11 @@ public:
    * @return const std::string& the upstream cluster that owns the route.
    */
   virtual const std::string& clusterName() const PURE;
+
+  /**
+   * @return const CorsPolicy* the CORS policy for this virtual host.
+   */
+  virtual const CorsPolicy* corsPolicy() const PURE;
 
   /**
    * Do potentially destructive header transforms on request headers prior to forwarding. For
@@ -261,6 +315,22 @@ public:
 };
 
 /**
+ * An interface representing the Decorator.
+ */
+class Decorator {
+public:
+  virtual ~Decorator() {}
+
+  /**
+   * This method decorates the supplied span.
+   * @param Tracing::Span& the span.
+   */
+  virtual void apply(Tracing::Span& span) const PURE;
+};
+
+typedef std::unique_ptr<const Decorator> DecoratorConstPtr;
+
+/**
  * An interface that holds a RedirectEntry or a RouteEntry for a request.
  */
 class Route {
@@ -276,6 +346,11 @@ public:
    * @return the route entry or nullptr if there is no matching route for the request.
    */
   virtual const RouteEntry* routeEntry() const PURE;
+
+  /**
+   * @return the decorator or nullptr if not defined for the request.
+   */
+  virtual const Decorator* decorator() const PURE;
 };
 
 typedef std::shared_ptr<const Route> RouteConstSharedPtr;