+++ source/common/http/async_client_impl.cc
@@ -11,6 +11,8 @@
 namespace Envoy {
 namespace Http {
 
+const std::list<std::string> AsyncStreamImpl::NullCorsPolicy::allow_origin_;
+const Optional<bool> AsyncStreamImpl::NullCorsPolicy::allow_credentials_;
 const std::vector<std::reference_wrapper<const Router::RateLimitPolicyEntry>>
     AsyncStreamImpl::NullRateLimitPolicy::rate_limit_policy_entry_;
 const AsyncStreamImpl::NullRateLimitPolicy AsyncStreamImpl::RouteEntryImpl::rate_limit_policy_;