+++ source/common/http/filter/fault_filter.cc
@@ -15,6 +15,7 @@
 #include "common/http/codes.h"
 #include "common/http/header_map_impl.h"
 #include "common/http/headers.h"
+#include "common/http/utility.h"
 #include "common/json/config_schemas.h"
 #include "common/router/config_impl.h"
 
@@ -171,7 +172,7 @@ Optional<uint64_t> FaultFilter::delayDuration() {
   return ret;
 }
 
-std::string FaultFilter::abortHttpStatus() {
+uint64_t FaultFilter::abortHttpStatus() {
   // TODO(mattklein123): check http status codes obtained from runtime.
   uint64_t http_status =
       config_->runtime().snapshot().getInteger(ABORT_HTTP_STATUS_KEY, config_->abortCode());
@@ -181,7 +182,7 @@ std::string FaultFilter::abortHttpStatus() {
         downstream_cluster_abort_http_status_key_, http_status);
   }
 
-  return std::to_string(http_status);
+  return http_status;
 }
 
 void FaultFilter::recordDelaysInjectedStats() {
@@ -228,7 +229,10 @@ FaultFilterStats FaultFilterConfig::generateStats(const std::string& prefix, Sta
   return {ALL_FAULT_FILTER_STATS(POOL_COUNTER_PREFIX(scope, final_prefix))};
 }
 
-void FaultFilter::onDestroy() { resetTimerState(); }
+void FaultFilter::onDestroy() {
+  resetTimerState();
+  stream_destroyed_ = true;
+}
 
 void FaultFilter::postDelayInjection() {
   resetTimerState();
@@ -243,12 +247,10 @@ void FaultFilter::postDelayInjection() {
 }
 
 void FaultFilter::abortWithHTTPStatus() {
-  // TODO(htuch): Switch this to Utility::sendLocalReply().
-  Http::HeaderMapPtr response_headers{
-      new HeaderMapImpl{{Headers::get().Status, abortHttpStatus()}}};
-  callbacks_->encodeHeaders(std::move(response_headers), true);
-  recordAbortsInjectedStats();
   callbacks_->requestInfo().setResponseFlag(Http::AccessLog::ResponseFlag::FaultInjected);
+  Http::Utility::sendLocalReply(*callbacks_, stream_destroyed_,
+                                static_cast<Http::Code>(abortHttpStatus()), "fault filter abort");
+  recordAbortsInjectedStats();
 }
 
 bool FaultFilter::matchesTargetUpstreamCluster() {