+++ source/common/router/config_impl.h
@@ -58,9 +58,36 @@ public:
   // Router::Route
   const RedirectEntry* redirectEntry() const override { return &SSL_REDIRECTOR; }
   const RouteEntry* routeEntry() const override { return nullptr; }
+  const Decorator* decorator() const override { return nullptr; }
 
 private:
   static const SslRedirector SSL_REDIRECTOR;
+};
+
+/**
+ * Implementation of CorsPolicy that reads from the proto route and virtual host config.
+ */
+class CorsPolicyImpl : public CorsPolicy {
+public:
+  CorsPolicyImpl(const envoy::api::v2::CorsPolicy& config);
+
+  // Router::CorsPolicy
+  const std::list<std::string>& allowOrigins() const override { return allow_origin_; };
+  const std::string& allowMethods() const override { return allow_methods_; };
+  const std::string& allowHeaders() const override { return allow_headers_; };
+  const std::string& exposeHeaders() const override { return expose_headers_; };
+  const std::string& maxAge() const override { return max_age_; };
+  const Optional<bool>& allowCredentials() const override { return allow_credentials_; };
+  bool enabled() const override { return enabled_; };
+
+private:
+  std::list<std::string> allow_origin_;
+  std::string allow_methods_;
+  std::string allow_headers_;
+  std::string expose_headers_;
+  std::string max_age_{};
+  Optional<bool> allow_credentials_{};
+  bool enabled_;
 };
 
 class ConfigImpl;
@@ -83,6 +110,7 @@ public:
   const ConfigImpl& globalRouteConfig() const { return global_route_config_; }
 
   // Router::VirtualHost
+  const CorsPolicy* corsPolicy() const override { return cors_policy_.get(); }
   const std::string& name() const override { return name_; }
   const RateLimitPolicy& rateLimitPolicy() const override { return rate_limit_policy_; }
 
@@ -115,6 +143,7 @@ private:
   std::vector<VirtualClusterEntry> virtual_clusters_;
   SslRequirements ssl_requirements_;
   const RateLimitPolicyImpl rate_limit_policy_;
+  std::unique_ptr<const CorsPolicyImpl> cors_policy_;
   const ConfigImpl& global_route_config_; // See note in RouteEntryImplBase::clusterEntry() on why
                                           // raw ref to the top level config is currently safe.
   std::list<std::pair<Http::LowerCaseString, std::string>> request_headers_to_add_;
@@ -173,6 +202,20 @@ private:
 };
 
 /**
+ * Implementation of Decorator that reads from the proto route decorator.
+ */
+class DecoratorImpl : public Decorator {
+public:
+  DecoratorImpl(const envoy::api::v2::Decorator& decorator);
+
+  // Decorator::apply
+  void apply(Tracing::Span& span) const override;
+
+private:
+  const std::string operation_;
+};
+
+/**
  * Base implementation for all route entries.
  */
 class RouteEntryImplBase : public RouteEntry,
@@ -195,6 +238,7 @@ public:
 
   // Router::RouteEntry
   const std::string& clusterName() const override;
+  const CorsPolicy* corsPolicy() const override { return cors_policy_.get(); }
   void finalizeRequestHeaders(Http::HeaderMap& headers) const override;
   const HashPolicy* hashPolicy() const override { return hash_policy_.get(); }
   Upstream::ResourcePriority priority() const override { return priority_; }
@@ -219,6 +263,7 @@ public:
   // Router::Route
   const RedirectEntry* redirectEntry() const override;
   const RouteEntry* routeEntry() const override;
+  const Decorator* decorator() const override { return decorator_.get(); }
 
 protected:
   const bool case_sensitive_;
@@ -247,6 +292,7 @@ private:
       return parent_->finalizeRequestHeaders(headers);
     }
 
+    const CorsPolicy* corsPolicy() const override { return parent_->corsPolicy(); }
     const HashPolicy* hashPolicy() const override { return parent_->hashPolicy(); }
     Upstream::ResourcePriority priority() const override { return parent_->priority(); }
     const RateLimitPolicy& rateLimitPolicy() const override { return parent_->rateLimitPolicy(); }
@@ -272,6 +318,7 @@ private:
     // Router::Route
     const RedirectEntry* redirectEntry() const override { return nullptr; }
     const RouteEntry* routeEntry() const override { return this; }
+    const Decorator* decorator() const override { return nullptr; }
 
   private:
     const RouteEntryImplBase* parent_;
@@ -310,9 +357,12 @@ private:
   static std::multimap<std::string, std::string>
   parseOpaqueConfig(const envoy::api::v2::Route& route);
 
+  static DecoratorConstPtr parseDecorator(const envoy::api::v2::Route& route);
+
   // Default timeout is 15s if nothing is specified in the route config.
   static const uint64_t DEFAULT_ROUTE_TIMEOUT_MS = 15000;
 
+  std::unique_ptr<const CorsPolicyImpl> cors_policy_;
   const VirtualHostImpl& vhost_; // See note in RouteEntryImplBase::clusterEntry() on why raw ref
                                  // to virtual host is currently safe.
   const bool auto_host_rewrite_;
@@ -335,6 +385,8 @@ private:
 
   // TODO(danielhochman): refactor multimap into unordered_map since JSON is unordered map.
   const std::multimap<std::string, std::string> opaque_config_;
+
+  const DecoratorConstPtr decorator_;
 };
 
 /**
@@ -371,6 +423,24 @@ public:
 
 private:
   const std::string path_;
+};
+
+/**
+ * Route entry implementation for regular expression match routing.
+ */
+class RegexRouteEntryImpl : public RouteEntryImplBase {
+public:
+  RegexRouteEntryImpl(const VirtualHostImpl& vhost, const envoy::api::v2::Route& route,
+                      Runtime::Loader& loader);
+
+  // Router::RouteEntry
+  void finalizeRequestHeaders(Http::HeaderMap& headers) const override;
+
+  // Router::Matchable
+  RouteConstSharedPtr matches(const Http::HeaderMap& headers, uint64_t random_value) const override;
+
+private:
+  const std::regex regex_;
 };
 
 /**